requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1

# Image Processing
Pillow==10.2.0

# File Upload Support
python-multipart==0.0.6

# Optional: For testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.26.0

# Optional: For production
# gunicorn==21.2.0
# redis==5.0.1
# python-jose[cryptography]==3.3.0  # For JWT authentication
# slowapi==0.1.9  # For rate limiting